converting code to strategy pattern
