+++ llm-server/routes/chat/implementation/functions_strategy.py
@@ -1,22 +1,20 @@
-from routes.chat.implementation.handler_interface import RequestHandler
-from typing import Callable, Dict, List, Any
+from routes.chat.implementation.handler_interface import ChatRequestHandler
+from typing import Dict, Optional
 
 from custom_types.response_dict import ResponseDict
-from routes.flow.utils.document_similarity_dto import DocumentSimilarityDTO
-from shared.models.opencopilot_db import ChatHistory
+from asyncio import Future
 
 
-class FunctionStrategy(RequestHandler):
-    def __init__(self, function: Callable):
-        self.function = function
-
+class FunctionStrategy(ChatRequestHandler):
     def handle_request(
         self,
-        top_documents: Dict[str, List[DocumentSimilarityDTO]],
-        chat_history: List[ChatHistory],
+        text: str,
         session_id: str,
-        user_message: str,
+        base_prompt: str,
+        bot_id: str,
+        headers: Dict[str, str],
+        app: Optional[str],
         is_streaming: bool,
-    ) -> ResponseDict:
+    ) -> Future[ResponseDict]:
         # Extract relevant information from inputs
         raise NotImplementedError("Subclasses must override handle_request.")
