+++ llm-server/routes/chat/implementation/handler_interface.py
@@ -1,18 +1,19 @@
 from abc import ABC, abstractmethod
-from typing import Dict, List
+from typing import Dict, Optional
 from custom_types.response_dict import ResponseDict
-from routes.flow.utils.document_similarity_dto import DocumentSimilarityDTO
-from shared.models.opencopilot_db import ChatHistory
+from asyncio import Future
 
 
-class RequestHandler(ABC):
+class ChatRequestHandler(ABC):
     @abstractmethod
-    def handle_request(
+    async def handle_request(
         self,
-        top_documents: Dict[str, List[DocumentSimilarityDTO]],
-        chat_history: List[ChatHistory],
+        text: str,
         session_id: str,
-        user_message: str,
+        base_prompt: str,
+        bot_id: str,
+        headers: Dict[str, str],
+        app: Optional[str],
         is_streaming: bool,
-    ) -> ResponseDict:
+    ) -> Future[ResponseDict]:
         pass
